wincng: make `ssh2_wcng_hash_update()` length `size_t` (was: `ULONG`)
Cherry-picked from #2171

+++ src/wincng.c
@@ -745,10 +745,10 @@ int ssh2_wcng_hash_init(struct wcng_hash_ctx *ctx, BCRYPT_ALG_HANDLE hAlg,
 }
 
 int ssh2_wcng_hash_update(struct wcng_hash_ctx *ctx,
-                          const void *data, ULONG datalen)
+                          const void *data, size_t datalen)
 {
-    int ret = BCryptHashData(ctx->hHash, SSH2_UNCONST(data), datalen, 0);
-
+    int ret = BCryptHashData(ctx->hHash,
+                             SSH2_UNCONST(data), (ULONG)datalen, 0);
     return BCRYPT_SUCCESS(ret);
 }
 
@@ -827,7 +827,7 @@ int ssh2_hmac_sha512_init(ssh2_hmac_ctx *ctx, void *key, size_t keylen)
 
 int ssh2_hmac_update(ssh2_hmac_ctx *ctx, const void *data, size_t datalen)
 {
-    return ssh2_wcng_hash_update(ctx, data, (ULONG)datalen);
+    return ssh2_wcng_hash_update(ctx, data, datalen);
 }
 
 int ssh2_hmac_final(ssh2_hmac_ctx *ctx, void *mac, size_t maclen)
